Use `libuuid` instead of `util-linux`
The `defaults` copy of the `util-linux` and the `conda-forge` copy of
the `libuuid` package overlap with each other. In fact, the `util-linux`
package only includes `libuuid`. As the `util-linux` package in
`defaults` is not maintained and saw its first and last build last year,
switch over to `libuuid`, which is maintained and serves the same
purpose.